projects/dev_agency_output/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic[email]==2.10.3
pyjwt==2.10.1
bcrypt==4.2.1
aiosqlite==0.20.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-anyio==0.0.0
